Get Typical Dishes of Colombia

Retrieves information about Colombia's typical dishes from the API Colombia. This function fetches only selected fields: id, name, description, and region.

Usage

get_colombia_typical_dishes()

Value

A tibble with multiple rows and the selected columns, or NULL if the API is unavailable.

Examples

# \donttest{
dishes <- get_colombia_typical_dishes()
print(dishes)
#> # A tibble: 68 × 3
#>       id name                  description                                      
#>    <int> <chr>                 <chr>                                            
#>  1     1 Bandeja Paisa         "La bandeja paisa es el orgullo de la cocina pai…
#>  2     2 Sancocho Antioqueño   "El sancocho antioqueño es un plato típico de An…
#>  3     3 Ajiaco Santafereño    " Sopa tradicional de Bogotá y el altiplano cund…
#>  4     4 Sancocho de Gallina   "Sopa tradicional del Valle del Cauca, ideal par…
#>  5     5 Lechona Tolimense     "Preparación típica para celebraciones, consiste…
#>  6     6 Arepas de Huevo       "Arepas fritas rellenas de huevo, típicas de la …
#>  7     7 Mazamorra Chiquita    "Sopa densa hecha a base de maíz, legumbres y ca…
#>  8     8 Sobrebarriga al Horno "Carne de res cocida al horno, acompañada con pa…
#>  9     9 Arroz Atollado        "Plato típico con arroz cocido mezclado con cerd…
#> 10    10 Tamal Tolimense       "Masa de maíz rellena de carne y vegetales, envu…
#> # ℹ 58 more rows
# }