Get Invasive Species of Colombia

Retrieves information about invasive species in Colombia from the API Colombia. This function fetches only selected fields: id, name, scientificName, commonNames, impact, manage, and riskLevel.

Usage

get_colombia_invasive_species()

Value

A tibble with multiple rows and the selected columns, or NULL if the API is unavailable.

Examples

# \donttest{
invasive_species <- get_colombia_invasive_species()
print(invasive_species)
#> # A tibble: 72 × 7
#>       id name                 scientificName commonNames impact manage riskLevel
#>    <int> <chr>                <chr>          <chr>       <chr>  <chr>      <int>
#>  1     1 Acacia negra, gris   Acacia decurr… "acacia ce… "Esta… A pes…         2
#>  2     2 Alfombra de agua     Azolla filicu… "Azola, he… "Esta… Primo…         2
#>  3     3 Algodón de seda      Calotropis pr… "algodón e… "Este… C. pr…         2
#>  4     4 Árbol del pan        Artocarpus al… "pandiaño,… "Como… Estos…         1
#>  5     5 Bambú de la India    Bambusa vulga… "bambú, gu… " B. … Se pu…         2
#>  6     6 Botón de oro         Tithonia dive… "botón de … "T. d… Cuand…         1
#>  7     7 Braquiaria           Urochloa decu… "brachiari… "este… Gener…         2
#>  8     8 Buchón               Limnobium lae… "buchón, b… "L. l… Lange…         2
#>  9    22 Helecho serrucho     Nephrolepis c… "helecho s… "N. c… se pu…         2
#> 10     9 Buchón de agua, tar… Eichhornia cr… "buchón, t… "es c… El co…         2
#> # ℹ 62 more rows
# }