Get Radios of Colombia

Retrieves information about radio stations in Colombia from the API Colombia. This function fetches only the following fields: id, name, frequency, and band.

Usage

get_colombia_radios()

Value

A tibble with multiple rows and the selected columns, or NULL if the API is unavailable.

Examples

# \donttest{
radios <- get_colombia_radios()
print(radios)
#> # A tibble: 41 × 4
#>       id name                   frequency band  
#>    <int> <chr>                      <dbl> <chr> 
#>  1    36 "La Reina"                  98.6 FM    
#>  2    33 "Barranquilla Estereo"      91.1 FM    
#>  3    31 "Fundinguie Radio"           0   Stream
#>  4    40 "La Cariñosa"              610   AM    
#>  5    11 "Tropicana FM"             103.  FM    
#>  6    10 "Vibra FM"                 105.  FM    
#>  7     9 "Los 40 Principales"        89.9 FM    
#>  8     8 "La FM"                     94.9 FM    
#>  9     7 "Radio uno"                 88.9 FM    
#> 10     6 "La Mega "                  90.9 FM    
#> # ℹ 31 more rows
# }